+++ src/arbiter/__main__.py
@@ -1,10 +1,13 @@
 import click
 import importlib.util
 import logging
+import sys
+import warnings
 
 from polyswarmartifact import ArtifactType
+from polyswarmclient import utils
 from polyswarmclient.config import init_logging
-from polyswarmclient.exceptions import FatalError
+from polyswarmclient.exceptions import FatalError, SecurityWarning
 
 logger = logging.getLogger(__name__)  # Initialize logger
 
@@ -48,8 +51,8 @@ def choose_backend(backend):
 @click.option('--log', default='WARNING', help='Logging level')
 @click.option('--client-log', default='WARNING',
               help='PolySwarm Client log level')
-@click.option('--polyswarmd-addr', envvar='POLYSWARMD_ADDR', default='api.polyswarm.network/v1/default',
-              help='Address (host:port) of polyswarmd instance')
+@click.option('--polyswarmd-addr', envvar='POLYSWARMD_ADDR', default='https://api.polyswarm.network/v1/default',
+              help='Address (scheme://host:port) of polyswarmd instance')
 @click.option('--keyfile', envvar='KEYFILE', type=click.Path(exists=True), default='keyfile',
               help='Keystore file containing the private key to use with this arbiter')
 @click.option('--password', envvar='PASSWORD', prompt=True, hide_input=True,
@@ -61,15 +64,17 @@ def choose_backend(backend):
 @click.option('--testing', default=0,
               help='Activate testing mode for integration testing, respond to N bounties then exit')
 @click.option('--insecure-transport', is_flag=True,
-              help='Connect to polyswarmd via http:// and ws://, mutually exclusive with --api-key')
+              help='Deprecated. Used only to change the default scheme to http in polyswarmd-addr if not present')
+@click.option('--allow-key-over-http', is_flag=True, envvar='ALLOW_KEY_OVER_HTTP',
+              help='Force api keys over http (Not Recommended)')
 @click.option('--chains', multiple=True, default=['side'],
               help='Chain(s) to operate on')
 @click.option('--log-format', default='text',
               help='Log format. Can be `json` or `text` (default)')
 @click.option('--artifact-type', multiple=True, default=['file'],
               help='List of artifact types to scan')
-def main(log, client_log, polyswarmd_addr, keyfile, password, api_key, backend, testing, insecure_transport, chains,
-         log_format, artifact_type):
+def main(log, client_log, polyswarmd_addr, keyfile, password, api_key, backend, testing, insecure_transport,
+         allow_key_over_http, chains, log_format, artifact_type):
     """
     Entrypoint for the arbiter driver
     """
@@ -84,16 +89,20 @@ def main(log, client_log, polyswarmd_addr, keyfile, password, api_key, backend, 
     init_logging(['arbiter', logger_name], log_format, loglevel)
     init_logging(['polyswarmclient'], log_format, clientlevel)
 
+    polyswarmd_addr = utils.finalize_polyswarmd_addr(polyswarmd_addr, api_key, allow_key_over_http, insecure_transport)
+    if insecure_transport:
+        warnings.warn('--insecure-transport will be removed soon. Please add http:// or https:// to polyswarmd-addr`',
+                      DeprecationWarning)
+
     artifact_types = None
     if artifact_type:
         artifact_types = [ArtifactType.from_string(artifact) for artifact in artifact_type]
 
     arbiter_class.connect(polyswarmd_addr, keyfile, password,
                           api_key=api_key, testing=testing,
-                          insecure_transport=insecure_transport,
                           chains=set(chains),
                           artifact_types=artifact_types).run()
 
 
 if __name__ == '__main__':
-    main()
+    main(sys.argv[1:])