+++ src/microengine/__main__.py
@@ -1,12 +1,14 @@
 import click
 import importlib.util
 import logging
+import sys
 import warnings
 
 from polyswarmartifact import ArtifactType
+from polyswarmclient import utils
 
 from polyswarmclient.config import init_logging, validate_apikey
-from polyswarmclient.exceptions import FatalError
+from polyswarmclient.exceptions import FatalError, SecurityWarning
 from polyswarmclient.filters.bountyfilter import split_filter, FilterComparison, BountyFilter
 from polyswarmclient.filters.confidencefilter import ConfidenceModifier
 from polyswarmclient.filters.filter import parse_filters
@@ -85,8 +87,8 @@ def choose_bid_strategy(bid_strategy):
               help='App Log level')
 @click.option('--client-log', default='WARNING',
               help='PolySwarm Client log level')
-@click.option('--polyswarmd-addr', envvar='POLYSWARMD_ADDR', default='api.polyswarm.network/v1/default',
-              help='Address (host:port) of polyswarmd instance')
+@click.option('--polyswarmd-addr', envvar='POLYSWARMD_ADDR', default='https://api.polyswarm.network/v1/default',
+              help='Address (scheme://host:port) of polyswarmd instance')
 @click.option('--keyfile', envvar='KEYFILE', type=click.Path(exists=True), default='keyfile',
               help='Keystore file containing the private key to use with this microengine')
 @click.option('--password', envvar='PASSWORD', prompt=True, hide_input=True,
@@ -99,7 +101,9 @@ def choose_bid_strategy(bid_strategy):
 @click.option('--testing', default=0,
               help='Activate testing mode for integration testing, respond to N bounties and N offers then exit')
 @click.option('--insecure-transport', is_flag=True,
-              help='Connect to polyswarmd via http:// and ws://, mutually exclusive with --api-key')
+              help='Deprecated. Used only to change the default scheme to http in polyswarmd-addr if not present')
+@click.option('--allow-key-over-http', is_flag=True, envvar='ALLOW_KEY_OVER_HTTP',
+              help='Force api keys over http (Not Recommended)')
 @click.option('--chains', multiple=True, default=['side'],
               help='Chain(s) to operate on')
 @click.option('--log-format', default='text',
@@ -131,17 +135,20 @@ def choose_bid_strategy(bid_strategy):
               ),
               help='Add filter in format `[favor|penalize] key [eq|gt|gte|lt|lte|startswith|endswith|regex] value` '
                    'to modify confidence based on metadata.')
-# @click.option('--offers', envvar='OFFERS', default=False, is_flag=True,
-#               help='Should the abassador send offers')
-def main(log, client_log, polyswarmd_addr, keyfile, password, api_key, backend, testing, insecure_transport, chains,
-         log_format, artifact_type, bid_strategy, accept, exclude, filter, confidence):
-    """Entrypoint for the microengine driver
+def main(log, client_log, polyswarmd_addr, keyfile, password, api_key, backend, testing, insecure_transport,
+         allow_key_over_http, chains, log_format, artifact_type, bid_strategy, accept, exclude, filter, confidence):
+    """ Entrypoint for the microengine driver
     """
     loglevel = getattr(logging, log.upper(), None)
     clientlevel = getattr(logging, client_log.upper(), None)
     if not isinstance(loglevel, int) or not isinstance(clientlevel, int):
         logging.error('invalid log level')
         raise FatalError('Invalid log level', 1)
+
+    polyswarmd_addr = utils.finalize_polyswarmd_addr(polyswarmd_addr, api_key, allow_key_over_http, insecure_transport)
+    if insecure_transport:
+        warnings.warn('--insecure-transport will be removed soon. Please add http:// or https:// to polyswarmd-addr`',
+                      DeprecationWarning)
 
     logger_name, microengine_class = choose_backend(backend)
     bid_logger_name, bid_strategy_class = choose_bid_strategy(bid_strategy)
@@ -171,9 +178,8 @@ def main(log, client_log, polyswarmd_addr, keyfile, password, api_key, backend, 
                               bounty_filter=BountyFilter(filter_accept, filter_reject),
                               chains=set(chains),
                               confidence_modifier=ConfidenceModifier(favor, penalize),
-                              insecure_transport=insecure_transport,
                               testing=testing).run()
 
 
 if __name__ == '__main__':
-    main()
+    main(sys.argv[1:])